+++ Sources/PieceMetalSuzuki/PieceMetalSuzuki.swift
@@ -254,31 +254,6 @@ func saveBufferToPng(buffer: CVPixelBuffer, format: CIFormat) -> Void {
     }
 }
 
-public func makeTextureFromCVPixelBuffer(
-    pixelBuffer: CVPixelBuffer, 
-    textureFormat: MTLPixelFormat,
-    textureCache: CVMetalTextureCache
-) -> MTLTexture? {
-    let width = CVPixelBufferGetWidth(pixelBuffer)
-    let height = CVPixelBufferGetHeight(pixelBuffer)
-    
-    // Create a Metal texture from the image buffer.
-    var cvTextureOut: CVMetalTexture?
-    let status = CVMetalTextureCacheCreateTextureFromImage(kCFAllocatorDefault, textureCache, pixelBuffer, nil, textureFormat, width, height, 0, &cvTextureOut)
-    guard status == kCVReturnSuccess else {
-        debugPrint("Error at CVMetalTextureCacheCreateTextureFromImage \(status)")
-        CVMetalTextureCacheFlush(textureCache, 0)
-        return nil
-    }
-    
-    guard let cvTexture = cvTextureOut, let texture = CVMetalTextureGetTexture(cvTexture) else {
-        CVMetalTextureCacheFlush(textureCache, 0)
-        return nil
-    }
-    
-    return texture
-}
-
 func loadChainStarterFunction(device: MTLDevice) -> MTLFunction? {
     do {
         guard let libUrl = Bundle.module.url(forResource: "PieceSuzukiKernel", withExtension: "metal", subdirectory: "Metal") else {
@@ -344,25 +319,4 @@ func createChainStarters(
     Profiler.add(end - start, to: .startChains)
     
     return true
-}
-
-extension MTLComputePipelineState {
-    func threadgroupParameters(texture: MTLTexture) -> (threadgroupsPerGrid: MTLSize, threadsPerThreadgroup: MTLSize) {
-        let threadHeight = maxTotalThreadsPerThreadgroup / threadExecutionWidth
-        return (
-            /// Subdivide grid as far as possible.
-            MTLSizeMake(threadExecutionWidth, threadHeight, 1),
-            /// Via https://developer.apple.com/documentation/metal/mtlcomputecommandencoder/1443138-dispatchthreadgroups
-            /// Also seen in https://developer.apple.com/documentation/avfoundation/additional_data_capture/avcamfilter_applying_filters_to_a_capture_stream
-            ///
-            /// The weird math is a form of rounding up using integer division.
-            /// If thread `width` or `height` evenly divides the texture `width` or `height`, that factor is returned.
-            /// Otherwise, the value is increased enough to push division to the next number, effectively rounding up.
-            MTLSizeMake(
-                (texture.width  + threadExecutionWidth - 1) / threadExecutionWidth,
-                (texture.height + threadHeight         - 1) / threadHeight,
-                1
-            )
-        )
-    }
 }
